+++ src/kernel.cpp
@@ -28,6 +28,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #include "process/task.h"
 #include "drivers/terminal.h"
 #include "shell/shell.h"
+#include "fs/vfs.h"
+#include "fs/ramdisk.h"
 
 #define THREAD_HZ 100
 
@@ -58,10 +60,13 @@ extern "C" void kernel_main(uint32_t magic, multiboot_info* mbi) {
     init_multitasking();
     init_timer(THREAD_HZ); // 100 Hz, i.e. every 10 ms
 
+    init_ramdisk();
+    vfs_mount(&ramdisk_ops);    // TODO: One day have a proper disk file system like EXT2 or FAT32 and mount onto it instead
+
     // Enable interrupts
     asm volatile("sti");
 
-    create_task(shell_thread);
+    create_task(shell_thread, "Shell");
 
     // The OS must NEVER die.
     // Interrupts take back control from this loop whenever